web/backend/requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
environ==1.0
jwt==1.3.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2